prepare the models
