+++ src/components/icons/index.ts
@@ -1,1 +1,3 @@
 export * as iconsRouter from './router.icons'
+export { IconInterface } from './interfaces.icons'
+export { getAllIcons } from './service.icons'
