+++ src/components/info/model.info.ts
